+++ include/Small/Core.hpp
@@ -1,0 +1,19 @@
+/**
+ * @file Core.hpp
+ * @author Alexandre Plateau (lexplt.dev@gmail.com)
+ * @brief Include all the files from the Core folder
+ * @version 0.1
+ * @date 2020-04-15
+ * 
+ * @copyright Copyright (c) 2020
+ * 
+ */
+
+#ifndef sgl_small_core
+#define sgl_small_core
+
+#include <Small/Core/Application.hpp>
+#include <Small/Core/SceneManager.hpp>
+#include <Small/Core/Scene.hpp>
+
+#endif