+++ include/Small/Graphics.hpp
@@ -1,0 +1,17 @@
+/**
+ * @file Graphics.hpp
+ * @author Alexandre Plateau (lexplt.dev@gmail.com)
+ * @brief Include all the files from the Graphics folder
+ * @version 0.1
+ * @date 2020-04-15
+ * 
+ * @copyright Copyright (c) 2020
+ * 
+ */
+
+#ifndef sgl_small_graphics
+#define sgl_small_graphics
+
+#include <Small/Graphics/Tilemap.hpp>
+
+#endif