+++ include/Small/System.hpp
@@ -1,0 +1,18 @@
+/**
+ * @file System.hpp
+ * @author Alexandre Plateau (lexplt.dev@gmail.com)
+ * @brief Include all the files from the System folder
+ * @version 0.1
+ * @date 2020-04-15
+ * 
+ * @copyright Copyright (c) 2020
+ * 
+ */
+
+#ifndef sgl_small_system
+#define sgl_small_system
+
+#include <Small/System/Platform.hpp>
+#include <Small/System/SharedLibrary.hpp>
+
+#endif